frontend/requirements.txt
api==0.0.7
hydralit_components==1.0.10
langchain==0.0.309
loguru==0.7.2
openai==0.27.8
openai==0.28.1
pandas==1.5.3
psutil==5.9.5
psycopg2_binary==2.9.6
pydantic==1.10.7
pydantic==1.10.13
python-dotenv==1.0.0
PyYAML==6.0
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
streamlit_authenticator==0.2.3
streamlit_elements==0.1.0
streamlit_extras==0.3.4
torch==2.0.1
torch==2.1.0
tqdm==4.66.1
transformers==4.33.3
transformers==4.34.0
weaviate_client==3.24.2
